Выполняет шаги из задания - Учебный проект: gotta get 'em all
