+++ public/css/style.css
@@ -29,6 +29,29 @@ body {
   font-family: "HelveticaNeueCyr", Arial, sans-serif;
 }
 
+@keyframes shake {
+  0%,
+  100% {
+    transform: translateX(0);
+  }
+  10%,
+  30%,
+  50%,
+  70%,
+  90% {
+    transform: translateX(-5px);
+  }
+  20%,
+  40%,
+  60%,
+  80% {
+    transform: translateX(5px);
+  }
+}
+.shake {
+  animation: shake 0.6s;
+}
+
 .container {
   width: 960px;
   padding: 0 20px;
@@ -191,6 +214,11 @@ body {
   background-color: #ffffff;
   outline: 0;
 }
+
+.card__inner--submit-error {
+  border: solid 1px red !important;
+}
+
 .card--deadline .card__inner {
   box-shadow: 0 2px 38px 0 rgba(240, 0, 0, 0.19);
 }
