+++ src/model-task.js
@@ -1,0 +1,21 @@
+export default class ModelTask {
+  constructor(data) {
+    this.id = data[`id`];
+    this.title = data[`title`] || ``;
+    this.dueDate = new Date(data[`due_date`]);
+    this.tags = new Set(data[`tags`] || []);
+    this.picture = data[`picture`] || ``;
+    this.repeatingDays = data[`repeating_days`];
+    this.color = data[`color`];
+    this.isFavorite = Boolean(data[`is_favorite`]);
+    this.isDone = Boolean(data[`is_done`]);
+  }
+
+  static parseTask(data) {
+    return new ModelTask(data);
+  }
+
+  static parseTasks(data) {
+    return data.map(ModelTask.parseTask);
+  }
+}
